requirements.txt
flask==3.0.3
python-telegram-bot==20.8
boto3==1.35.0
python-dotenv==1.0.1
